requirements.txt
tweepy
uvicorn
starlette
pyyaml
